tools/ARL-NPoC/requirements.txt
urllib3==1.25.3
requests==2.26.0
lxml>=4.3.0
colorlog==6.7.0
requests_ntlm==1.1.0
pycryptodomex==3.16.0
pymongo==3.13.0
dnslib==0.9.16
Flask==2.0.3
PyYAML==5.4.1